\name{BioCore.Annotation-package}
\alias{BioCore.Annotation-package}
\alias{BioCore.Annotation}
\docType{package}
\title{
BioCore Annotation R Package
}
\description{
A R package used in the BioCore group for annotation and filtering of ANNOVAR output data.
}
\tabular{11}{
Package: \tab BioCore.Annotation\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-01-1\cr
License: \tab GPL-2\cr
}
\author{
Richard J. de Borja <richard.deborja@uhnresearch.ca>
}
\keyword{annotation}
