\name{create_variant_signature_barplot}
\alias{create_variant_signature_barplot}
\title{
Create the variant signature barplot
}
\description{
Create a stacked barplot illustrating each of the 6 possible pyrimidine
based change (i.e. C>A, C>G, C>T, T>A, T>C, T>G).
}
\usage{
create_variant_signature_barplot(data=NULL)
}
\arguments{
  \item{data}{dataframe containing the pyrimidine based change}
}
\value{
Returns a ggplot2 plot object.
}
\author{
Richard J. de Borja
}
\keyword{ signature }
\keyword{ snv}
\keyword{ variant }
